Split CLI commands into modules
